+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Parser
   - Make `ASTNode.type/start/text/children` readonly.
+  - Add `ASTNode.traverse` to traverse AST after parsing.
   - ELR Parser
     - Export `ELR.commit` and `ELR.rollback`.
     - **Breaking Change**: Reducer will take `ReducerContext` as the parameter instead of `ParserContext`.