+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.5.1
+
+- Optimize package size.
+
 ## v0.5.0
 
 - Parser